% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadc.R
\name{bb_handler_aadc}
\alias{bb_handler_aadc}
\title{Handler for files downloaded from the Australian Antarctic Data Centre EDS system}
\usage{
bb_handler_aadc(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_wget}}}
}
\value{
TRUE on success
}
\description{
This is a handler function to be used with data from the Australian Antarctic Data Centre. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}. AADC EDS files have a URL of the form https://data.aad.gov.au/eds/file/wxyz/ or https://data.aad.gov.au/eds/wxyz/download where wxyz is a numeric file identifier.
}
\references{
http://data.aad.gov.au
}
