% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amps_handler.R
\name{bb_handler_amps}
\alias{bb_handler_amps}
\title{Handler for AMPS data (Antarctic mesoscale prediction system)}
\usage{
bb_handler_amps(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_wget}}}
}
\value{
TRUE on success
}
\description{
This is a handler function to be used with AMPS data from http://www2.mmm.ucar.edu/rt/amps/. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
}
\references{
http://www2.mmm.ucar.edu/rt/amps/
}
