% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarview_handler.R
\name{bb_handler_polarview}
\alias{bb_handler_polarview}
\title{Handler for Polarview Sentinel-1 data}
\usage{
bb_handler_polarview(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link[bowerbird:bb_rget]{bowerbird::bb_rget()}}}
}
\value{
\code{TRUE} on success
}
\description{
This is a handler function to be used with Sentinel-1 data from \url{https://polarview.aq}. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link[bowerbird:bb_source]{bowerbird::bb_source()}}.
}
\references{
\url{https://polarview.aq}
}
