% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natice_handler.R
\name{bb_handler_usnic}
\alias{bb_handler_usnic}
\title{Handler for US National Ice Center charts}
\usage{
bb_handler_usnic(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_rget}}}
}
\value{
TRUE on success
}
\description{
This is a handler function to be used with US National Ice Center charts from https://usicecenter.gov/Products/AntarcHome. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
}
\details{
Note that the USNIC server does not support timestamp operations on requests, so it is not possible to download only files that have changed since last downloaded. Bowerbird configurations with \code{clobber = 1} (download if modified) will be treated as \code{clobber = 0} (don't download if file already exists).

This handler can take a \code{method} argument as specified in the \code{\link[bowerbird]{bb_source}} constructor:
\itemize{
  \item chart_type string: either "filled" [default] or "vector"
}
}
\references{
https://usicecenter.gov/Products/AntarcHome
}
