% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarview.R
\name{bb_polarview_search}
\alias{bb_polarview_search}
\title{Search the PolarView catalogue}
\usage{
bb_polarview_search(
  acquisition_date = Sys.Date() + -14:0,
  formats = c("jpg", "geotiff"),
  polygon = NULL,
  verbose = FALSE
)
}
\arguments{
\item{acquisition_date}{Date: the allowable image acquisition dates}

\item{formats}{character: one or more of "jpg" (jpg preview images) "jp2" or "geotiff". Note that the geotiffs are much larger than the jpg previews}

\item{polygon}{string or sfc_POLYGON: either an \code{sfc_POLYGON} or a string giving a polygon in WKT format and EPSG:3031 projection. Only images intersecting this polygon will be returned}

\item{verbose}{logical: if \code{TRUE}, show additional progress output}
}
\value{
A character vector of image URLs
}
\description{
This function is used by \code{\link{bb_handler_polarview}}. Users probably won't need to use it directly.
}
\references{
<https://www.polarview.aq>
}
