% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{bb_aadc_source}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bowerbird}{\code{\link[bowerbird]{bb_aadc_source}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

