% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources}
\alias{sources}
\title{Bowerbird configurations for various Antarctic and Southern Ocean data sources}
\usage{
sources(name, formats, time_resolutions, ...)
}
\arguments{
\item{name}{character vector: only return data sources with name or id matching these values}

\item{formats}{character: for some sources, the format can be specified. See thematic source functions for details}

\item{time_resolutions}{character: for some sources, the time resolution can be specified. See thematic source functions for details}

\item{...}{: other parameters passed to thematic source functions}
}
\value{
A tibble
}
\description{
The \code{sources} function is a convenience wrapper around the thematic functions: \code{\link[=sources_seaice]{sources_seaice()}}, \code{\link[=sources_altimetry]{sources_altimetry()}}, etc.
}
\references{
See \code{reference} and \code{citation} field in each row of the returned tibble
}
