% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_meteorological.R
\name{sources_meteorological}
\alias{sources_meteorological}
\title{Reanalysis data sources}
\usage{
sources_meteorological(name, formats, time_resolutions, ...)
}
\arguments{
\item{name}{character vector: only return data sources with name or id matching these values}

\item{formats}{character: for some sources, the format can be specified. See the list of sources above for details}

\item{time_resolutions}{character: for some sources, the time resolution can be specified. See the list of sources above for details}

\item{...}{: additional source-specific parameters. See the list of sources above for details}
}
\value{
a tibble with columns as specified by \code{\link{bb_source}}
}
\description{
Data sources providing meteorological data.
}
\details{
\itemize{
  \item "Antarctic Mesoscale Prediction System grib files": The Antarctic Mesoscale Prediction System - AMPS - is an experimental, real-time numerical weather prediction capability that provides support for the United States Antarctic Program, Antarctic science, and international Antarctic efforts
}

The returned tibble contains more information about each source.
}
\examples{
\dontrun{
## define a configuration and add the AMPS data to it
cf <- bb_config("/my/file/root") \%>\%
  bb_add(sources_meteorological("Antarctic Mesoscale Prediction System grib files"))
}
}
\references{
See the \code{doc_url} and \code{citation} field in each row of the returned tibble for references associated with these particular data sources
}
\seealso{
\code{\link{sources_altimetry}}, \code{\link{sources_ocean_colour}}, \code{\link{sources_oceanographic}}, \code{\link{sources_reanalysis}}, \code{\link{sources_seaice}}, \code{\link{sources_sst}}, \code{\link{sources_topography}}
}
