% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_oceanographic.R
\name{sources_oceanographic}
\alias{sources_oceanographic}
\title{Oceanographic data sources}
\usage{
sources_oceanographic(name, formats, time_resolutions, ...)
}
\arguments{
\item{name}{character vector: only return data sources with name or id matching these values}

\item{formats}{character: for some sources, the format can be specified. See the list of sources above for details}

\item{time_resolutions}{character: for some sources, the time resolution can be specified. See the list of sources above for details}

\item{...}{: additional source-specific parameters. See the list of sources above for details}
}
\value{
a tibble with columns as specified by \code{\link{bb_source}}
}
\description{
Data sources providing oceanographic data.
}
\details{
\itemize{
  \item "CSIRO Atlas of Regional Seas 2009": CARS is a digital climatology, or atlas of seasonal ocean water properties
  \item "World Ocean Atlas 2009": World Ocean Atlas 2009 is included here for convenience but has been superseded by the World Ocean Atlas 2013 V2
  \item "World Ocean Atlas 2013 V2": World Ocean Atlas 2013 version 2 (WOA13 V2) is a set of objectively analyzed (1 degree grid) climatological fields of in situ temperature, salinity, dissolved oxygen, Apparent Oxygen Utilization (AOU), percent oxygen saturation, phosphate, silicate, and nitrate at standard depth levels for annual, seasonal, and monthly compositing periods for the World Ocean. It also includes associated statistical fields of observed oceanographic profile data interpolated to standard depth levels on 5 degree, 1 degree, and 0.25 degree grids
}

The returned tibble contains more information about each source.
}
\examples{
\dontrun{
## define a configuration and add the Atlas of Regional Seas data to it
cf <- bb_config("/my/file/root")
src <- sources_oceanographic("CSIRO Atlas of Regional Seas 2009")
cf <- bb_add(cf,src)
}

}
\references{
See the \code{doc_url} and \code{citation} field in each row of the returned tibble for references associated with these particular data sources
}
\seealso{
\code{\link{sources_altimetry}}, \code{\link{sources_meteorological}}, \code{\link{sources_ocean_colour}}, \code{\link{sources_reanalysis}}, \code{\link{sources_seaice}}, \code{\link{sources_sst}}, \code{\link{sources_topography}}
}
