% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_reanalysis.R
\name{sources_reanalysis}
\alias{sources_reanalysis}
\title{Reanalysis data sources}
\usage{
sources_reanalysis(name, formats, time_resolutions, ...)
}
\arguments{
\item{name}{character vector: only return data sources with name or id matching these values}

\item{formats}{character: for some sources, the format can be specified. See the list of sources above for details}

\item{time_resolutions}{character: for some sources, the time resolution can be specified. See the list of sources above for details}

\item{...}{: additional source-specific parameters. See the list of sources above for details}
}
\value{
a tibble with columns as specified by \code{\link{bb_source}}
}
\description{
Data sources providing data from global reanalysis models.
}
\details{
\itemize{
  \item "NCEP-DOE Reanalysis 2": NCEP-DOE Reanalysis 2 is an improved version of the NCEP Reanalysis I model that fixed errors and updated paramterizations of of physical processes. Accepts \code{time_resolution} values of "6 hour", "day", and/or "month" (default). The 6-hourly data is the original output time resolution. Daily and monthly averages are calculated from the 6-hourly model output
  \item "NCEP/NCAR Reanalysis 1": The NCEP/NCAR Reanalysis 1 project is using a state-of-the-art analysis/forecast system to perform data assimilation using past data from 1948 to the present. Only the monthly data are so far included here
}

The returned tibble contains more information about each source.
}
\examples{
\dontrun{
## define a configuration and add the monthly NCEP2 data to it
cf <- bb_config("/my/file/root") \%>\%
  bb_add(sources_reanalysis("NCEP-DOE Reanalysis 2",time_resolution="month"))
}
}
\references{
See the \code{doc_url} and \code{citation} field in each row of the returned tibble for references associated with these particular data sources
}
\seealso{
\code{\link{sources_altimetry}}, \code{\link{sources_biological}}, \code{\link{sources_meteorological}}, \code{\link{sources_ocean_colour}}, \code{\link{sources_oceanographic}}, \code{\link{sources_sdm}}, \code{\link{sources_seaice}}, \code{\link{sources_sst}}, \code{\link{sources_topography}}
}
