% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_sdm.R
\name{sources_sdm}
\alias{sources_sdm}
\title{Data sources intended for species distribution modelling and similar tasks}
\usage{
sources_sdm(name, formats, time_resolutions, ...)
}
\arguments{
\item{name}{character vector: only return data sources with name or id matching these values}

\item{formats}{character: for some sources, the format can be specified. See the list of sources above for details}

\item{time_resolutions}{character: for some sources, the time resolution can be specified. See the list of sources above for details}

\item{...}{: additional source-specific parameters. See the list of sources above for details}
}
\value{
a tibble with columns as specified by \code{\link[bowerbird:bb_source]{bowerbird::bb_source()}}
}
\description{
Data sources providing environmental and similar gridded data, suitable for species distribution modelling, regionalisation analyses, and similar tasks.
}
\details{
\itemize{
\item "Southern Ocean marine environmental data": a collection of gridded marine environmental data layers suitable for use in Southern Ocean species distribution modelling. All environmental layers have been generated at a spatial resolution of 0.1 degrees, covering the Southern Ocean extent (80 degrees S - 45 degrees S, -180 - 180 degrees). The layers include information relating to bathymetry, sea ice, ocean currents, primary production, particulate organic carbon, and other oceanographic data. See the vignette for more information: \code{vignette("SO_SDM_data", package = "blueant")}
}

The returned tibble contains more information about each source.
}
\examples{
\dontrun{
## define a configuration, storing data in a temporary directory
cf <- bb_config(local_file_root = tempdir())

## add the marine environmental data layers
cf <- cf \%>\% bb_add(sources_sdm("Southern Ocean marine environmental data"))

## sync it (get the data)
res <- bb_sync(cf, verbose = TRUE)

## see the vignette for more information on this data source:
vignette("SO_SDM_data", package = "blueant")
}
}
\references{
See the \code{doc_url} and \code{citation} field in each row of the returned tibble for references associated with these particular data sources
}
