% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amps_handler.R
\name{bb_handler_amps}
\alias{bb_handler_amps}
\title{Handler for AMPS data (Antarctic mesoscale prediction system)}
\usage{
bb_handler_amps(config, verbose = FALSE, local_dir_only = FALSE)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}

\item{verbose}{logical: if TRUE, provide additional progress output}

\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
}
\value{
the directory if local_dir_only is TRUE, otherwise TRUE on success
}
\description{
Handler for AMPS data (Antarctic mesoscale prediction system)
}
\references{
http://www2.mmm.ucar.edu/rt/amps/
}
