% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo_handler.R
\name{bb_handler_argo}
\alias{bb_handler_argo}
\title{Handler for Argo data sources}
\usage{
bb_handler_argo(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_rget}}}
}
\value{
TRUE on success
}
\description{
This is a handler function to be used with Argo data. Tested with the ifremer data centre, not yet tested with others. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
}
\references{
https://wwz.ifremer.fr/en/Research-Technology/Scientific-departments/Department-of-Marine-and-Digital-Infrastructures/The-French-ARGO-Data-Centre
}
