% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueant.R
\docType{package}
\name{blueant}
\alias{blueant}
\title{\pkg{blueant}}
\description{
A collection of data source definitions that can be used with the bowerbird package. These sources define a range of environmental and other data sources useful to Antarctic and Southern Ocean studies.
}
