% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{blueant_sources}
\alias{blueant_sources}
\title{Bowerbird configurations for various Antarctic and Southern Ocean data sources}
\usage{
blueant_sources(name, data_group)
}
\arguments{
\item{name}{character vector: only return data sources with name or id matching these values}

\item{data_group}{character vector: only return data sources belonging to these data groups}
}
\value{
tibble
}
\description{
Bowerbird configurations for various Antarctic and Southern Ocean data sources
}
\examples{
blueant_sources()

}
\references{
See \code{reference} and \code{citation} field in each row of the returned tibble
}
\seealso{
\code{\link{bb_config}}
}
